logica_python/requirements.txt
python-docx
pyhanko
fastapi
uvicorn
pyHanko
docx2pdf
python-multipart
cryptography
nest_asyncio
## Para instalar necesitas usar pip install -r requirements.txt
